fix: טיפול בהערות ריוויו — ולידציה, ריכוז סכמות ושרידות OTP
- ולידציה ל-SwitchStationRequest: station_id חייב להיות חיובי
- ריכוז סכמות מולטי-תחנה ב-panel/schemas.py (מקור אחד)
- שרידות OTP: try/except סביב store_otp עם לוג ותגובה ידידותית
  כשל Redis לא ישאיר את המשתמש ללא תגובה ברורה
